refactor: PRレビュー指摘に対応（単一クエリ化・一括追記・テスト共有化）
- findByStepId を application リレーション経由の 1 クエリに統合（2往復→1往復・
  承認Inbox 最頻読取の改善／レビュー#1）
- update の終端イベント永続化を for ループ upsert から createMany skipDuplicates へ。
  テーブル毎1往復に畳み version ロック保持中の往復数を削減。既決は @id 衝突で
  スキップされ created_at を保持（方式A の冪等性は維持・レビュー#5）
- テストの cleanup を cleanupApprovalFixtures(numbers) として helper に共有化。
  免除版が steps/イベント削除を欠く部分集合だった問題を上位集合で解消（レビュー#4）
- translateInsertConflict / update のコメントを実装と整合（一律P2002翻訳が安全な
  根拠を明記、count0 は append-only ゆえ version不一致のみと明記・レビュー#2/#9）

enum キャスト（#8）は既存 EstimateMapper の house style のため本PRでは据え置く。
